% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spray.R
\name{add_spray}
\alias{add_spray}
\title{Add a foliar treated spray to the state variables}
\usage{
add_spray(.y, .time, .year, parms)
}
\arguments{
\item{.y}{A vector with the current state variables}

\item{.time}{A numeric giving the current time}

\item{.year}{An integer giving the current year}

\item{parms}{A list containing the simulation parameters}
}
\value{
A vector with the updated state variables
}
\description{
This function adds the relevant dose to the state variables
Management information is in the `parms$management` data.frame
}
