% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_effective_life.R
\name{calculate_effective_life}
\alias{calculate_effective_life}
\title{Calculate the effective life}
\usage{
calculate_effective_life(df, max_HAD, threshold = 10)
}
\arguments{
\item{df}{A data.frame as returned from `simulate()`}

\item{max_HAD}{Numeric, the maximum HAD in a simulation with no pathogen}

\item{threshold}{An optional parameter, giving the threshold acceptable HAD loss as a percentage}
}
\value{
A vector
}
\description{
Calculate the effective life
}
