% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_HAD.R
\name{calculate_HAD}
\alias{calculate_HAD}
\title{Calculate the healthy area duration (HAD) of each field each year}
\usage{
calculate_HAD(df)
}
\arguments{
\item{df}{The data.frame returned from simulate()}
}
\value{
A data.frame giving the HAD in each field for each year
}
\description{
Calculate the healthy area duration (HAD) of each field each year
}
