% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_resistance_frequency.R
\name{calculate_resistance_frequency}
\alias{calculate_resistance_frequency}
\title{Calculate resistance frequency for each gene in each field of the given data.frame}
\usage{
calculate_resistance_frequency(df)
}
\arguments{
\item{df}{A data.frame in which the first two columns are the time and year, and then the state variables for each field in turn}
}
\value{
A data.frame with a column for the resistance frequency of each gene in each field. Columns are labelled: "rf_field_'fieldID'_gene_'geneID'".
}
\description{
Calculate resistance frequency for each gene in each field of the given data.frame
}
\examples{
parameters() |> simulate() |> calculate_resistance_frequency()
}
