% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_SLIR.R
\name{calculate_SLIR}
\alias{calculate_SLIR}
\title{Calculate S, L, I, R leaf area index for each field}
\usage{
calculate_SLIR(sim, year = NA)
}
\arguments{
\item{sim}{A data.frame, as produced by `simulate()`}

\item{year}{(optional) an integer, specifying a single year to extract}
}
\value{
A data.frame containing the values of each variable in each field in the model
}
\description{
Calculate S, L, I, R leaf area index for each field
}
\examples{
simulate(parameters()) |>
    calculate_SLIR(year=2)
}
