% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_strain_frequencies.R
\name{calculate_strain_frequency}
\alias{calculate_strain_frequency}
\title{Calculate the frequency of each strain in each field of the given data.frame}
\usage{
calculate_strain_frequency(df)
}
\arguments{
\item{df}{A data.frame in which the first two columns are the time and year, and then the state variables for each field in turn}
}
\value{
A data.frame with a column for the frequency of each strain in each field. Columns are labelled: "rf_field_'fieldID'_gene_'geneID'".
}
\description{
Calculate the frequency of each strain in each field of the given data.frame
}
\examples{
parameters(n_years=1) |>
   simulate() |>
   calculate_strain_frequency()
}
