% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parameters.R
\name{check_parameters}
\alias{check_parameters}
\title{Check the integrity of the passed parameters}
\usage{
check_parameters(parameters)
}
\arguments{
\item{parameters}{A list of parameters, as passed to `simulate()`}
}
\description{
This function checks that the dimensions of the parameters passed to `simulate()`
}
