% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_management.R
\name{create_management}
\alias{create_management}
\title{Create a management data.frame}
\usage{
create_management(field1 = NULL, field2 = NULL, field3 = NULL, n_years)
}
\arguments{
\item{field1}{The management in field 1: a data.frame with three columns; Fungicide, Dose, Time}

\item{field2}{The management in field 2: A data.frame with three columns; Fungicide, Dose, Time}

\item{field3}{The management in field 3: A data.frame with three columns; Fungicide, Dose, Time}

\item{n_years}{An integer, the number of years to run a simulation}
}
\value{
A data.frame, containing all the management events over all the years
}
\description{
Create a management data.frame
}
\examples{
f1 = data.frame(
  Fungicide = c(1,1,3),
  Dose = 1,
  Time = c(0,400,700)
)
f2 = data.frame(
  Fungicide = c(2,1,2,3),
  Dose = c(1,0.5,0.5,1),
  Time = c(0,400,400,700)
)
f3 = data.frame(
  Fungicide = c(3,3,3),
  Dose = 1,
  Time = c(0,400,700)
)
mngmt = create_management(field1 = f1, field2 = f2, field3 = f3,n_years = 2)
}
