% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_growth.R
\name{crop_growth}
\alias{crop_growth}
\title{Calculate the derivative of crop growth}
\usage{
crop_growth(.t, max_area_index, growth_rate, midpoint)
}
\arguments{
\item{.t}{The current time (in degree days)}

\item{max_area_index}{The maximum area index}

\item{growth_rate}{Growth rate when density is zero}

\item{midpoint}{Midpoint of the crop growth logistic curve (in degree days)}
}
\value{
A double giving the derivative of the crop area index at time t
}
\description{
The following function calculates the rate of crop growth
We are assuming that the canopy dynamics are not affected by the pathogen
}
