% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_senescence.R
\name{crop_senescence}
\alias{crop_senescence}
\title{Calculate the rate of crop senescence at time t}
\usage{
crop_senescence(.t, max_area_index, growth_rate, midpoint)
}
\arguments{
\item{.t}{Time}

\item{max_area_index}{Maximum area index}

\item{growth_rate}{Crop growth rate}

\item{midpoint}{Midpoint of the senescence curve}
}
\value{
A numeric
}
\description{
Calculate the rate of crop senescence at time t
}
