% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incoming_spores.R
\name{dispersal}
\alias{dispersal}
\title{Calculate the rate of spore influx from all other fields}
\usage{
dispersal(
  y,
  iField,
  dispersal_proportion,
  n_genotypes,
  transmission_rate,
  field_size,
  n_fields
)
}
\arguments{
\item{y}{The state variables}

\item{iField}{The field that we're calculating the spore influx to}

\item{dispersal_proportion}{The proportion of spores that leave each field}

\item{n_genotypes}{The number of genotypes}

\item{transmission_rate}{The transmission rate}

\item{field_size}{A vector, the size of each field}

\item{n_fields}{The number of fields}
}
\value{
A numeric containing the total number of spores entering field 'iField'
}
\description{
Calculate the rate of spore influx from all other fields
}
