% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_model.R
\name{full_model}
\alias{full_model}
\title{Function passed to ode, returning the derivatives of all state variables}
\usage{
full_model(t, y, parms, year, genotype_proportions)
}
\arguments{
\item{t}{Numeric, the current time}

\item{y}{The current state variables}

\item{parms}{A list containing all the model parameters}

\item{year}{Integer, giving the year}

\item{genotype_proportions}{Genotype proportions of the initial inoculum for this year}
}
\value{
A list, whose first element is a vector of the derivatives
}
\description{
The parameters are those needed for the ode solver, deSolve::ode()
}
