% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungicide_effect.R
\name{fungicide_effect}
\alias{fungicide_effect}
\title{Return the proportional effect of a fungicide on a given genotype}
\usage{
fungicide_effect(.dose, .genotype, .fung_index, dra, drs)
}
\arguments{
\item{.dose}{Current dose}

\item{.genotype}{Current genotype}

\item{.fung_index}{Index of fungicide}

\item{dra}{Matrix, giving all the dose-response curve asymptotes}

\item{drs}{dose-response curve shapes}
}
\value{
A numeric giving the value of the dose-response curve
}
\description{
Return the proportional effect of a fungicide on a given genotype
}
