% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_frequencies.R
\name{genotype_frequencies}
\alias{genotype_frequencies}
\title{Calculate the frequency of each genotype in each field at the start of a year}
\usage{
genotype_frequencies(stubble, parms)
}
\arguments{
\item{stubble}{A list of matrices, of length `year_i-1`, containing the infectious leaf area at the end of each year}

\item{parms}{List of parameters}
}
\value{
A matrix, each row of which corresponds to a field. The resistance frequency of the primary inoculum for the coming year
}
\description{
Calculate the frequency of each genotype in each field at the start of a year
}
