% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables}
\alias{get_variables}
\title{Extract requested variables from a data.frame formatted according to `simulate()`}
\usage{
get_variables(.data, elements, field_index = NULL)
}
\arguments{
\item{.data}{A data.frame, of the same format of that returned by `simulate()`}

\item{elements}{A character vector containing one or more of the following: "S", "L", "I", "R", "D"}

\item{field_index}{(optional) The index of the field to extract}
}
\value{
A data.frame
}
\description{
Extract requested variables from a data.frame formatted according to `simulate()`
}
