% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_frequencies.R
\name{initial_geno_freq}
\alias{initial_geno_freq}
\title{Calculate the genotype frequency in year 1}
\usage{
initial_geno_freq(parms)
}
\arguments{
\item{parms}{The parameters object, as produced by `parameters()`}
}
\value{
A matrix containing the frequency of each genotype (column) for each field (row)
}
\description{
Calculate the genotype frequency in year 1
}
