% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_variables.R
\name{initialise_variables}
\alias{initialise_variables}
\title{Create a vector of all the variables in the model, the initial parameters for the ode simulator}
\usage{
initialise_variables(parms)
}
\arguments{
\item{parms}{List of parameters}
}
\value{
A numeric vector, giving the area index S,L_G,I_G,R,F_F for each field at the start of a growing season
}
\description{
Create a vector of all the variables in the model, the initial parameters for the ode simulator
}
