% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_latent_period.R
\name{latent_period}
\alias{latent_period}
\title{Calculate the latent period at a given time, based on the fungicide dose}
\usage{
latent_period(
  fungicide_doses,
  n_genotypes,
  n_fungicides,
  latent_period,
  dose_response_asymptote,
  dose_response_shape
)
}
\arguments{
\item{fungicide_doses}{Vector of numeric. The dose of each fungicide}

\item{n_genotypes}{Integer, the number of genotypes}

\item{n_fungicides}{Integer, the number of fungicides}

\item{latent_period}{Numeric, the latent period}

\item{dose_response_asymptote}{The asymptotes of the dose-response curves}

\item{dose_response_shape}{The shapes of the dose-response curves}
}
\value{
A vector giving the latent period for each genotype
}
\description{
Calculate the latent period at a given time, based on the fungicide dose
}
