% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mutation_matrix.R
\name{mutation_matrix}
\alias{mutation_matrix}
\title{Create a matrix, giving the probability of mutating from genotype i to genotype j}
\usage{
mutation_matrix(n_genotypes, mutation_rate, n_resistance_genes)
}
\arguments{
\item{n_genotypes}{Integer, the number of genotypes}

\item{mutation_rate}{The mutation rate}

\item{n_resistance_genes}{The number of resistance genes}
}
\value{
A matrix of n_genotypes^2
}
\description{
Create a matrix, giving the probability of mutating from genotype i to genotype j
}
