% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{mutations}
\alias{mutations}
\title{Calculate the amount of mutation in secondary inoculum}
\usage{
mutations(inoculum, n_resistance_genes, n_genotypes, mutation_rate)
}
\arguments{
\item{inoculum}{Current newly produced secondary inoculum}

\item{n_resistance_genes}{Integer, the number of resistance genes}

\item{n_genotypes}{Integer, number of genotypes}

\item{mutation_rate}{Numeric, the mutation rate}
}
\value{
The amount of secondary inoculum of each genotype
}
\description{
Calculate the amount of mutation in secondary inoculum
}
