% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Create a parameters object.}
\usage{
parameters(
  crop_growth_midpoint = 550,
  crop_growth_rate = 0.01,
  crop_max_area_index = 2.75,
  crop_senesence_midpoint = 1050,
  transmission_rate = 0.1172795,
  transmission_fitness_cost = 0,
  latent_period = 225,
  infectious_period = 225,
  sporulation_rate = 0.4001843,
  initial_resistance_frequency = rep(0, 2),
  inoculum_scale = 40,
  inoculum_shape = 1,
  mutation_rate = 1e-07,
  overwinter_dispersal_proportion = 1,
  spore_escape_ratio = 0.1,
  stubble_survival = 0.1,
  sex_prop = 0,
  psi = 0.5,
  field_size = c(0.5, 0.25, 0.25),
  initial_inoculum = rep(0.1304277, 3),
  dose_response_asymptote = matrix(c(0.68813, 0.8563, 0.68813, 0.8563, 0.8604, 0.8604,
    0.6, 0.6, 0.6, 0.6, 0.6, 0.6), ncol = 4, nrow = 3, byrow = TRUE),
  dose_response_shape = matrix(c(6.26876, 0.255, 6.26876, 0.255, 27.88, 27.88, 2.4, 2.4,
    4, 4, 4, 4), ncol = 4, nrow = 3, byrow = TRUE),
  decay_rate = c(0.006, 0.00513, 0.006),
  seed_treatment_coefficient = c(0.248778, 0.7984493, 0.961949),
  seed_treatment_scale = 20,
  seed_treatment_shape = 4,
  management = NULL,
  n_fields = 3,
  n_fungicides = 3,
  n_resistance_genes = 2,
  n_years = 20,
  punnet_square = matrix(NA, ncol = 4, nrow = 4),
  max_time = 1200
)
}
\arguments{
\item{crop_growth_midpoint}{The midpoint of the growth curve, in degree days since crop emergence}

\item{crop_growth_rate}{The growth rate of the crop canopy (leaf area per degree day)}

\item{crop_max_area_index}{The maximum area index of the crop canopy (leaf area per unit ground area)}

\item{crop_senesence_midpoint}{The midpoint of crop senescence}

\item{transmission_rate}{The transmission rate of the pathogen}

\item{transmission_fitness_cost}{The fitness cost of the transmission rate, between 0 and 1}

\item{latent_period}{Latent period of the pathogen}

\item{infectious_period}{Infectious period of the pathogen}

\item{sporulation_rate}{The sporulation rate}

\item{initial_resistance_frequency}{The initial resistance frequency, a vector of length n_resistance_genes}

\item{inoculum_scale}{A numeric, the scale parameter of inoculum release}

\item{inoculum_shape}{A numeric, the shape parameter of inoculum release}

\item{mutation_rate}{A numeric, between 0 and 1, the mutation rate}

\item{overwinter_dispersal_proportion}{The proportion of spores that overwinter}

\item{spore_escape_ratio}{The proportion of spores that escape}

\item{stubble_survival}{The proportion of stubble survival}

\item{sex_prop}{# Proportion of overwinter spores that go through sexual recombination}

\item{psi}{# Linkage disequilibirum coefficient}

\item{field_size}{A vector, containing the size of each field in the landscape}

\item{initial_inoculum}{A vector of size 3 specifying the amount of inoculum at the start of each season in each field}

\item{dose_response_asymptote}{A matrix, containing the asymptote of the dose-response relationship for each fungicide:genotype pair. The number of rows of the matrix should be the number of fungicides, and the number of columns should equal the number of genotypes}

\item{dose_response_shape}{A matrix, containing the shape of the dose-response relationship for each fungicide:genotype pair. The number of rows of the matrix should be the number of fungicides, and the number of columns should equal the number of genotypes}

\item{decay_rate}{A vector, containing the decay rate of each fungicide}

\item{seed_treatment_coefficient}{A vector with an element for each fungicide, containing the coefficient of the seed treatment for each fungicide}

\item{seed_treatment_scale}{A numeric, being the scale parameter of the seed treatment release for all fungicides}

\item{seed_treatment_shape}{A numeric, being the shape parameter of the seed treatment release for all fungicides}

\item{management}{(optional) A data.frame specifying when fungicide is applied to each year and each field. Required columns are: Year, Field, Fungicide, Dose, Time}

\item{n_fields}{An integer, giving the number of fields}

\item{n_fungicides}{An integer, specifying the number of fungicides}

\item{n_resistance_genes}{An integer, specifying the number of resistance genes}

\item{n_years}{An integer, specifying the number of years to simulate}

\item{punnet_square}{An empty matrix}

\item{max_time}{An integer, specifying the maximum time (in degree days) to simulate within each growing season}
}
\value{
A list, containing parameter values to be passed to `simulate()`
}
\description{
Creates a list containing all the parameters that can be set to run the NetBlotch model.
}
