% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_inoculum.R
\name{primary_inoculum}
\alias{primary_inoculum}
\title{The rate of primary inoculum influx at time t}
\usage{
primary_inoculum(
  .time,
  field_index,
  initial_inoculum,
  inoculum_shape,
  inoculum_scale,
  genotype_proportions
)
}
\arguments{
\item{.time}{The current time}

\item{field_index}{Integer, specifying which of the N fields in the landscape we're calculating for}

\item{initial_inoculum}{A vector containing the amount of initial inoculum in each field}

\item{inoculum_shape}{The shape parameter of the primary inoculum curve}

\item{inoculum_scale}{The scale parameter of the primary inoculum curve}

\item{genotype_proportions}{The proportion of each of the genotypes in the simulation at the start of the season}
}
\value{
A numeric giving the total influx of primary inoculum
}
\description{
The rate of primary inoculum influx at time t
}
