% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_treatment_deriv.R
\name{seed_treatment_deriv}
\alias{seed_treatment_deriv}
\title{Calculate the influx of dose from seed treatment at time t}
\usage{
seed_treatment_deriv(
  .year,
  .field,
  .t,
  .fungicide,
  management,
  seed_treatment_shape,
  seed_treatment_scale,
  seed_treatment_coefficient
)
}
\arguments{
\item{.year}{A numeric, giving the current year. Needed to get the correct spray}

\item{.field}{An integer giving the field number}

\item{.t}{A double giving the current time}

\item{.fungicide}{An integer specifying the fungicide}

\item{management}{data.frame giving the management details}

\item{seed_treatment_shape}{Shape of seed treatment}

\item{seed_treatment_scale}{Scale of seed treatment}

\item{seed_treatment_coefficient}{Coefficient of seed treatment}
}
\value{
A single double, giving the derivative for a single fungicide in field .field at time .t
}
\description{
This function returns the rate of influx of seed treatment at time t
}
