% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate the model}
\usage{
simulate(parms)
}
\arguments{
\item{parms}{A list containing all the parameters for the model, created by `parameters()`}
}
\value{
A data.frame, with the first column being the year, and second column time, followed by the state variables
}
\description{
Simulate the model
}
