% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_one_year.R
\name{simulate_one_year}
\alias{simulate_one_year}
\title{Simulate a single year}
\usage{
simulate_one_year(year_i, parms, stubble)
}
\arguments{
\item{year_i}{Integer giving the current year of the simulation (starting from 1)}

\item{parms}{The list of parameters in the model}

\item{stubble}{A list containing the genotype frequencies from the stubble at the end of each year}
}
\value{
A data.frame
}
\description{
Simulate a single year
}
