% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transmission_rate.R
\name{transmission_rate}
\alias{transmission_rate}
\title{Calculate the transmission rate at time t in field i, given the current fungicide dose}
\usage{
transmission_rate(
  fungicide_doses,
  transmission_rate,
  transmission_rate_fitness_cost,
  n_genotypes,
  n_fungicides,
  dose_response_asymptote,
  dose_response_shape
)
}
\arguments{
\item{fungicide_doses}{A vector giving the dose of each fungicide in field i}

\item{transmission_rate}{The default transmission rate}

\item{transmission_rate_fitness_cost}{A fitness cost associated with the transmission rate. Values of 0.0 indicate no fitness cost, while values of 1.0 indicate a full fitness cost.}

\item{n_genotypes}{Integer, the number of genotypes}

\item{n_fungicides}{Integer, the number of fungicides}

\item{dose_response_asymptote}{Matrix, the asymptotes of the dose-response curves}

\item{dose_response_shape}{Matrix, the shapes of the dose-response curves}
}
\value{
A vector giving the transmission rate of each genotype
}
\description{
Calculate the transmission rate at time t in field i, given the current fungicide dose
}
