% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variables}
\alias{variables}
\title{Return the variables for a single field}
\usage{
variables(y, field_index)
}
\arguments{
\item{y}{The whole vector of all variables for all fields}

\item{field_index}{An integer giving the number of the desired field}
}
\value{
A double vector with the variables for the single field
}
\description{
Return the variables for a single field
}
